+++ lib/middleware/Auth/ApiKey.ts
@@ -3,10 +3,15 @@ import {isFn} from '@valkyriestudios/utils/function';
 import {isNeString} from '@valkyriestudios/utils/string';
 import {
     Sym_TriFrostDescription,
+    Sym_TriFrostFingerPrint,
     Sym_TriFrostName,
     Sym_TriFrostType,
 } from '../../types/constants';
 import {type TriFrostContext} from '../../types/context';
+import {Sym_TriFrostMiddlewareAuth} from './types';
+
+/* Specific symbol attached to auth mware to identify them by */
+export const Sym_TriFrostMiddlewareApiKeyAuth = Symbol('TriFrost.Middleware.ApiKeyAuth');
 
 export type ApiKeyAuthResult = {key:string};
 
@@ -70,6 +75,8 @@ export function ApiKeyAuth <
     Reflect.set(mware, Sym_TriFrostName, 'TriFrostApiKeyAuth');
     Reflect.set(mware, Sym_TriFrostType, 'middleware');
     Reflect.set(mware, Sym_TriFrostDescription, 'API Key Authentication middleware');
+    Reflect.set(mware, Sym_TriFrostMiddlewareAuth, true);
+    Reflect.set(mware, Sym_TriFrostFingerPrint, Sym_TriFrostMiddlewareApiKeyAuth);
 
     return mware;
 }