+++ lib/middleware/Auth/Bearer.ts
@@ -3,10 +3,15 @@ import {isFn} from '@valkyriestudios/utils/function';
 import {isString} from '@valkyriestudios/utils/string';
 import {
     Sym_TriFrostDescription,
+    Sym_TriFrostFingerPrint,
     Sym_TriFrostName,
     Sym_TriFrostType,
 } from '../../types/constants';
 import {type TriFrostContext} from '../../types/context';
+import {Sym_TriFrostMiddlewareAuth} from './types';
+
+/* Specific symbol attached to auth mware to identify them by */
+export const Sym_TriFrostMiddlewareBearerAuth = Symbol('TriFrost.Middleware.BearerAuth');
 
 export type BearerAuthResult = {token:string};
 
@@ -57,6 +62,8 @@ export function BearerAuth <
     Reflect.set(mware, Sym_TriFrostName, 'TriFrostBearerAuth');
     Reflect.set(mware, Sym_TriFrostType, 'middleware');
     Reflect.set(mware, Sym_TriFrostDescription, 'HTTP Bearer Token Authentication middleware');
+    Reflect.set(mware, Sym_TriFrostMiddlewareAuth, true);
+    Reflect.set(mware, Sym_TriFrostFingerPrint, Sym_TriFrostMiddlewareBearerAuth);
 
     return mware;
 }