+++ lib/middleware/Security.ts
@@ -4,10 +4,14 @@ import {isIntGt} from '@valkyriestudios/utils/number';
 import {isNeArray} from '@valkyriestudios/utils/array';
 import {
     Sym_TriFrostDescription,
+    Sym_TriFrostFingerPrint,
     Sym_TriFrostName,
     Sym_TriFrostType,
 } from '../types/constants';
 import {type TriFrostContext} from '../types/context';
+
+/* Specific symbol attached to security mware to identify them by */
+export const Sym_TriFrostMiddlewareSecurity = Symbol('TriFrost.Middleware.Security');
 
 export enum ContentSecurityPolicy {
     DefaultSrc      = 'default-src',
@@ -510,6 +514,7 @@ export function Security (opts:TriFrostSecurityOptions = {}) {
     Reflect.set(mware, Sym_TriFrostName, 'TriFrostSecurity');
     Reflect.set(mware, Sym_TriFrostType, 'middleware');
     Reflect.set(mware, Sym_TriFrostDescription, 'Middleware for configuring Security headers and CSP on contexts passing through it');
+    Reflect.set(mware, Sym_TriFrostFingerPrint, Sym_TriFrostMiddlewareSecurity);
 
     return mware;
 }