+++ lib/middleware/index.ts
@@ -1,4 +1,13 @@
-export *                from './Auth';
-export {CacheControl}   from './CacheControl';
-export {Cors}           from './Cors';
-export {Security}       from './Security';
+export * from './Auth';
+export {
+    CacheControl,
+    Sym_TriFrostMiddlewareCacheControl
+} from './CacheControl';
+export {
+    Cors,
+    Sym_TriFrostMiddlewareCors
+} from './Cors';
+export {
+    Security,
+    Sym_TriFrostMiddlewareSecurity
+} from './Security';